latest from master

+++ includes/abstracts/abstract-wc-rest-controller.php
@@ -93,6 +93,7 @@ abstract class WC_REST_Controller extends WP_REST_Controller {
 		}
 
 		if ( $total > $limit ) {
+			/* translators: %s: items limit */
 			return new WP_Error( 'woocommerce_rest_request_entity_too_large', sprintf( __( 'Unable to accept more than %s items for this request.', 'woocommerce' ), $limit ), array( 'status' => 413 ) );
 		}
 
@@ -167,6 +168,12 @@ abstract class WC_REST_Controller extends WP_REST_Controller {
 
 		if ( ! empty( $items['delete'] ) ) {
 			foreach ( $items['delete'] as $id ) {
+				$id = (int) $id;
+
+				if ( 0 === $id ) {
+					continue;
+				}
+
 				$_item = new WP_REST_Request( 'DELETE' );
 				$_item->set_query_params( array( 'id' => $id, 'force' => true ) );
 				$_response = $this->delete_item( $_item );
@@ -188,7 +195,7 @@ abstract class WC_REST_Controller extends WP_REST_Controller {
 	/**
 	 * Validate a text value for a text based setting.
 	 *
-	 * @since 2.7.0
+	 * @since 3.0.0
 	 * @param string $value
 	 * @param array  $setting
 	 * @return string
@@ -202,7 +209,7 @@ abstract class WC_REST_Controller extends WP_REST_Controller {
 	/**
 	 * Validate select based settings.
 	 *
-	 * @since 2.7.0
+	 * @since 3.0.0
 	 * @param string $value
 	 * @param array  $setting
 	 * @return string|WP_Error
@@ -218,7 +225,7 @@ abstract class WC_REST_Controller extends WP_REST_Controller {
 	/**
 	 * Validate multiselect based settings.
 	 *
-	 * @since 2.7.0
+	 * @since 3.0.0
 	 * @param array $values
 	 * @param array  $setting
 	 * @return string|WP_Error
@@ -245,7 +252,7 @@ abstract class WC_REST_Controller extends WP_REST_Controller {
 	/**
 	 * Validate image_width based settings.
 	 *
-	 * @since 2.7.0
+	 * @since 3.0.0
 	 * @param array $value
 	 * @param array $setting
 	 * @return string|WP_Error
@@ -271,7 +278,7 @@ abstract class WC_REST_Controller extends WP_REST_Controller {
 	/**
 	 * Validate radio based settings.
 	 *
-	 * @since 2.7.0
+	 * @since 3.0.0
 	 * @param string $value
 	 * @param array  $setting
 	 * @return string|WP_Error
@@ -283,7 +290,7 @@ abstract class WC_REST_Controller extends WP_REST_Controller {
 	/**
 	 * Validate checkbox based settings.
 	 *
-	 * @since 2.7.0
+	 * @since 3.0.0
 	 * @param string $value
 	 * @param array  $setting
 	 * @return string|WP_Error
@@ -302,7 +309,7 @@ abstract class WC_REST_Controller extends WP_REST_Controller {
 	/**
 	 * Validate textarea based settings.
 	 *
-	 * @since 2.7.0
+	 * @since 3.0.0
 	 * @param string $value
 	 * @param array  $setting
 	 * @return string
@@ -322,7 +329,7 @@ abstract class WC_REST_Controller extends WP_REST_Controller {
 	/**
 	 * Add meta query.
 	 *
-	 * @since 2.7.0
+	 * @since 3.0.0
 	 * @param array $args       Query args.
 	 * @param array $meta_query Meta query.
 	 * @return array
@@ -352,16 +359,25 @@ abstract class WC_REST_Controller extends WP_REST_Controller {
 					'description' => __( 'List of created resources.', 'woocommerce' ),
 					'type'        => 'array',
 					'context'     => array( 'view', 'edit' ),
+					'items'       => array(
+						'type'    => 'object',
+					),
 				),
 				'update' => array(
 					'description' => __( 'List of updated resources.', 'woocommerce' ),
 					'type'        => 'array',
 					'context'     => array( 'view', 'edit' ),
+					'items'       => array(
+						'type'    => 'object',
+					),
 				),
 				'delete' => array(
 					'description' => __( 'List of delete resources.', 'woocommerce' ),
 					'type'        => 'array',
 					'context'     => array( 'view', 'edit' ),
+					'items'       => array(
+						'type'    => 'integer',
+					),
 				),
 			),
 		);
